+++ CHANGELOG.md
@@ -8,13 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### ToDo
-- static filters for Search::show
 - consider how to group/collapse the same cve's !!!!! if not tree then dialog window
 - onetomany or manytomany for wazych alert -> ticket
 - migrations: install is not upgrade
-- check all defaults columns for search views.
 - check wazuh version
-- wazuh tab: when connection problem then no results at all
 - assets tab, show trash problems. Not when search::show ????
 
 ## [0.0.3]
@@ -24,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ticket urgency level based on average points of severity level items attached to
 - partial API fetch for long data
 - fetching only data older than latest time discovered ?
+- static filters for Search::show
+- wazuh tab: connection problem fix
+
 
 ## [0.0.2] 2025-03-31